+++ src/test/kotlin/com/yapp/demo/DemoApplicationTests.kt
@@ -1,6 +1,6 @@
 package com.yapp.demo
 
-import com.yapp.demo.auth.external.apple.ApplePrivateKey
+import com.yapp.demo.auth.internal.ApplePrivateKeyLoader
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.bean.override.mockito.MockitoBean
@@ -8,7 +8,7 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean
 @SpringBootTest
 class DemoApplicationTests {
     @MockitoBean
-    private lateinit var applePrivateKey: ApplePrivateKey
+    private lateinit var applePrivateKeyLoader: ApplePrivateKeyLoader
 
     @Test
     fun contextLoads() {
